+++ flatlaf-intellij-themes/src/main/java/com/formdev/flatlaf/intellijthemes/FlatGradiantoDarkFuchsiaIJTheme.java
@@ -39,14 +39,6 @@ public class FlatGradiantoDarkFuchsiaIJTheme
 		}
 	}
 
-	/**
-	 * @deprecated use {@link #setup()} instead; this method will be removed in a future version
-	 */
-	@Deprecated
-	public static boolean install() {
-		return setup();
-	}
-
 	public static void installLafInfo() {
 		installLafInfo( NAME, FlatGradiantoDarkFuchsiaIJTheme.class );
 	}