+++ flatlaf-intellij-themes/src/main/java/com/formdev/flatlaf/intellijthemes/FlatNordIJTheme.java
@@ -39,14 +39,6 @@ public class FlatNordIJTheme
 		}
 	}
 
-	/**
-	 * @deprecated use {@link #setup()} instead; this method will be removed in a future version
-	 */
-	@Deprecated
-	public static boolean install() {
-		return setup();
-	}
-
 	public static void installLafInfo() {
 		installLafInfo( NAME, FlatNordIJTheme.class );
 	}